# docs/master/swagger.yaml
basePath: /api/v1/master
definitions:
  ErrorMessages:
    properties:
      message:
        type: string
      name:
        type: string
      path:
        type: string
      type:
        type: string
      validator:
        type: string
      value: {}
    type: object
  Response:
    properties:
      code:
        type: integer
      count:
        type: integer
      data: {}
      error_code:
        type: string
      error_messages:
        items:
          $ref: '#/definitions/ErrorMessages'
        type: array
      message:
        type: string
      page:
        type: integer
      status:
        type: boolean
      total:
        type: integer
      total_page:
        type: integer
    type: object
  model.Device:
    properties:
      additional:
        type: string
      code:
        description: Public identifier for simulators
        type: string
      description:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
    type: object
  model.DeviceAPI:
    properties:
      code:
        description: Public identifier for simulators
        type: string
      description:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
    type: object
  model.DeviceUpdateRequest:
    properties:
      code:
        type: string
      description:
        type: string
      location:
        type: string
      name:
        type: string
    type: object
  model.Sensor:
    properties:
      additional:
        type: string
      category:
        $ref: '#/definitions/model.SensorCategory'
      category_id:
        type: string
      device:
        $ref: '#/definitions/model.Device'
      device_id:
        type: string
      name:
        type: string
      readings:
        items:
          $ref: '#/definitions/model.SensorReading'
        type: array
      unit:
        type: string
    type: object
  model.SensorAPI:
    properties:
      category_id:
        type: string
      device_id:
        type: string
      name:
        type: string
      unit:
        type: string
    type: object
  model.SensorCategory:
    properties:
      additional:
        type: string
      code:
        type: string
      default_unit:
        type: string
      description:
        type: string
      name:
        type: string
      units:
        type: string
    required:
    - code
    - default_unit
    - name
    type: object
  model.SensorCategoryRequest:
    properties:
      code:
        type: string
      default_unit:
        type: string
      description:
        type: string
      name:
        type: string
      units:
        items:
          type: string
        type: array
    type: object
  model.SensorIngestRequest:
    properties:
      sensor_id:
        type: string
      timestamp:
        type: string
      value: {}
    required:
    - sensor_id
    - value
    type: object
  model.SensorReading:
    properties:
      created_at:
        type: string
      id:
        type: string
      retry_count:
        type: integer
      sensor_id:
        type: string
      status:
        type: string
      timestamp:
        type: string
      updated_at:
        type: string
      value: {}
    type: object
  model.SensorUpdateRequest:
    properties:
      name:
        type: string
      type:
        type: string
      unit:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: armadamuhammads@gmail.com
    name: Armada Muhammad Siswanto
  description: API Documentation
  title: API
  version: 1.0.0
paths:
  /devices:
    get:
      consumes:
      - application/json
      description: Get all devices
      produces:
      - application/json
      responses:
        "200":
          description: List of Devices
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Device'
                  type: array
              type: object
      security:
      - TokenKey: []
      summary: Get all devices
      tags:
      - Device
    post:
      consumes:
      - application/json
      description: Create a new device
      parameters:
      - description: Body Request
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/model.DeviceAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Device data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Device'
              type: object
      security:
      - TokenKey: []
      summary: Create a new device
      tags:
      - Device
  /devices/{device_id}/sensors:
    get:
      consumes:
      - application/json
      description: Get all sensors for a device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Sensors
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Sensor'
                  type: array
              type: object
      security:
      - TokenKey: []
      summary: Get all sensors for a device
      tags:
      - Sensor
    post:
      consumes:
      - application/json
      description: Create a new sensor for a device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Body Request
        in: body
        name: sensor
        required: true
        schema:
          $ref: '#/definitions/model.SensorAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Sensor data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Sensor'
              type: object
      security:
      - TokenKey: []
      summary: Create a new sensor for a device
      tags:
      - Sensor
  /devices/{device_id}/sensors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sensor
      parameters:
      - description: Sensor ID
        in: path
        name: id
        required: true
        type: string
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/Response'
      security:
      - TokenKey: []
      summary: Delete a sensor
      tags:
      - Sensor
    get:
      consumes:
      - application/json
      description: Get a sensor by id
      parameters:
      - description: Sensor ID
        in: path
        name: id
        required: true
        type: string
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sensor data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Sensor'
              type: object
      security:
      - TokenKey: []
      summary: Get a sensor by id
      tags:
      - Sensor
    put:
      consumes:
      - application/json
      description: Update a sensor
      parameters:
      - description: Sensor ID
        in: path
        name: id
        required: true
        type: string
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Body Request
        in: body
        name: sensor
        required: true
        schema:
          $ref: '#/definitions/model.SensorUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sensor data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Sensor'
              type: object
      security:
      - TokenKey: []
      summary: Update a sensor
      tags:
      - Sensor
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/Response'
      security:
      - TokenKey: []
      summary: Delete a device
      tags:
      - Device
    get:
      consumes:
      - application/json
      description: Get a device by ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Device'
              type: object
      security:
      - TokenKey: []
      summary: Get a device by ID
      tags:
      - Device
    put:
      consumes:
      - application/json
      description: Update a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Body Request
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/model.DeviceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Device data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Device'
              type: object
      security:
      - TokenKey: []
      summary: Update a device
      tags:
      - Device
  /sensor-categories:
    get:
      consumes:
      - application/json
      description: Get all sensor categories
      produces:
      - application/json
      responses:
        "200":
          description: List of Sensor Categories
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SensorCategory'
                  type: array
              type: object
      security:
      - TokenKey: []
      summary: Get all sensor categories
      tags:
      - SensorCategory
    post:
      consumes:
      - application/json
      description: Create a new sensor category
      parameters:
      - description: Body Request
        in: body
        name: sensor_category
        required: true
        schema:
          $ref: '#/definitions/model.SensorCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Sensor Category data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SensorCategory'
              type: object
      security:
      - TokenKey: []
      summary: Create a new sensor category
      tags:
      - SensorCategory
  /sensor-categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sensor category
      parameters:
      - description: Sensor Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/Response'
      security:
      - TokenKey: []
      summary: Delete a sensor category
      tags:
      - SensorCategory
    get:
      consumes:
      - application/json
      description: Get a sensor category by id
      parameters:
      - description: Sensor Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sensor Category data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SensorCategory'
              type: object
      security:
      - TokenKey: []
      summary: Get a sensor category by id
      tags:
      - SensorCategory
    put:
      consumes:
      - application/json
      description: Update a sensor category
      parameters:
      - description: Sensor Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Body Request
        in: body
        name: sensor_category
        required: true
        schema:
          $ref: '#/definitions/model.SensorCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sensor Category data
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SensorCategory'
              type: object
      security:
      - TokenKey: []
      summary: Update a sensor category
      tags:
      - SensorCategory
  /sensors/ingest:
    post:
      consumes:
      - application/json
      description: Ingest a sensor reading
      parameters:
      - description: Sensor reading data
        in: body
        name: sensor_reading
        required: true
        schema:
          $ref: '#/definitions/model.SensorIngestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/Response'
      security:
      - TokenKey: []
      summary: Ingest a sensor reading
      tags:
      - Sensor Ingest
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  TokenKey:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"